requirements.txt
allure_python_commons
playwright
pytest
pytest_playwright
pytest-html-reporter
pur
allure-pytest
python-dotenv
requests